remove warning on windows

+++ src/main.cpp
@@ -35,8 +35,8 @@ extern void render_next_gl(uint64_t time_micros)
 {
         double const phase = 6.30 * time_micros / 1e6 / 6.0;
         float sincos[2] = {
-                0.49f * sinf(phase),
-                0.49f * cosf(phase),
+                static_cast<float>(0.49 * sin(phase)),
+                static_cast<float>(0.49 * cos(phase)),
         };
         float const argb[4] = {
                 0.0f, 0.31f + 0.09f * sincos[0], 0.27f + 0.09f * sincos[1], 0.29f
